// src/enums/studentEnums.ts
export enum ProductStatus {
    ACTIVE = 1,
    INACTIVE = 0,
}

export enum PaymentStatus {
    PAID = "Paid",
    Pending = "Pending",
    REJECTED = "Rejected",
}

export enum AcceptEnquiryStatus {
    ACTIVE = 0,
    INACTIVE = 1,
}

export enum CourseStatus {
    ACTIVE = 1,
    INACTIVE = 0,
}

export enum CourseAcceptEnquiryStatus {
    ACTIVE = 0,
    INACTIVE = 1,
}

export enum PaymentFrequency {
    MONTHLY = "Monthly",
    QUATERLY = "Quaterly",
    HALF_YEARLY = "Half-Yearly",
    YEARLY = "Yearly",
    LUM_SUM = "Lum-sum",
}